Directory cleanup - new files for existing funcs
